Improve Nintendo Ware format handling and reading
- Add BRSTM, BCSTM, and BFSTM file reading.
- Fix a rare case where an integer overflow would result in bad GC-ADPCM encoding.
- Rename Bxstm namespace to NintendoWare because it now includes Bxwav formats as well.
- Add Bxwav 010 Editor binary templates.
- Read the new value (possibly a checksum) from BFSTM version 0.5.0.0 and up.
- Refactor NintendoWare namespace to more closely follow the structure of Nintendo Ware file formats.

+++ src/VGAudio/Codecs/GcAdpcm/GcAdpcmEncoder.cs
@@ -159,7 +159,7 @@ namespace VGAudio.Codecs.GcAdpcm
                     // Clamp and store
                     pcmOut[s + 2] = Clamp16(scaledSample);
                     // Accumulate distance
-                    int actualDistance = pcmIn[s + 2] - pcmOut[s + 2];
+                    double actualDistance = pcmIn[s + 2] - pcmOut[s + 2];
                     totalDistance += actualDistance * actualDistance;
                 }
